// src/types/HealthData.ts
export type HealthData = {
  steps: number;
  height: number;
  weight: number;  
  sleepDuration: number;
  distance: number;
  calories: number;
  menstruationFlow: string | null;
  menstruationPeriod: string | null;
  heartRate: string | number | null;
  hydration: number | null;
  bloodPressure: {
    systolic: number | null;
    diastolic: number | null;
  } | null;
  bodyTemperature: number | null;
  activeCalories: number | null;
  basalBodyTemperature: number | null;
  basalMetabolicRate: number | null;
  bloodGlucose: number | null;
  bodyFat: number | null;
  boneMass: number | null;
  cervicalMucus: string | null;
  elevationGained: number | null;
  exerciseSession: string | null;
  floorsClimbed: number | null;
  leanBodyMass: number | null;
  nutrition: string | null;
  ovulationTest: string | null;
  oxygenSaturation: number | null;
  power: number | null;
  respiratoryRate: number | null;
  restingHeartRate: number | null;
  sexualActivity: boolean | null;
  sexualActivityTimestamp: string | null;
  speed: number | null;
  stepsCadence: number | null;
  vo2Max: number | null;
  wheelchairPushes: number | null;
};